bindings: Work for const dictionary
IDL dictionary is a kind of container, and its const getter APIs should
return const members as a base rule.
We would like to follow the rule in the new code generator as much as
possible, and this CL updates some code to work for the coming changes.

But the change has no visible behavior changes.


Bug: 839389
Change-Id: I3fac82f2a9e5b7e0d1f4a0bd05481ea763af38c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2121595
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768682}