Pass on the used color-scheme to popup documents.
We set the used color-scheme on the owner element for select, date
pop-ups, etc. as the preferred color-scheme for the document used to
render the popup so that we can utilize @media queries to implement
light/dark theme in the UA.

Additionally added a null check to fix 1082389.

Bug: 929098, 1082389
Change-Id: Ifec2296e5a04d25e1837186b5d264172129fb041
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196170
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768702}