[mpc] Handle dependencies in the MatchedPropertiesCache
A "dependency" for a given CachedMatchedProperties object is defined
as a CSS property, whose computed value must be equal for the cached
parent style and the incoming parent style in order for a given
cache lookup to be considered a cache hit.

This CL makes it possible to add such a list of dependencies to
a CachedMatchedProperties object, but doesn't actually use this
capability yet (except from unit tests).

Bug: 1057072
Change-Id: I37cf3b88b6e09d2503865c5d7d9c2c7b2539d13f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2193682
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768737}