[conversions] Add impression / conversion storage limits
This CL adds an initial limit of 1024 impressions / conversions per site
to prevent excessive disk usage. In a follow-up we will consider
migrating this logic to call out to the quota manager.

This CL also deletes the MockStorageDelegate in favor of using the one
in conversion_test_utils.

Bug: 1075564
Change-Id: I4097432b39f1751cf96c6c75cf3830fbd75a2900
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167533
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768814}