Move V8 callback execution from TestRunnerForSpecificView.
Add methods in TestRunnerBindings to wrap a V8 function into a
base::Callback, making the function and any bound V8 arguments into
v8::Persistent objects.

The resulting callback, when Run(), can receive further V8 arguments,
which will appended to the bound v8::Persistent objects (converted back
to v8::Local objects). Then the v8 function will be invoked. This is
effectively currying for the v8 function arguments. Sorry, not sorry.

Also add a helper to PostTask a V8 function. It will be used as more
code migrates from TestRunnerForSpecificView.

We replace all calls that were being made to TestRunnerForSpecificView
from TestRunnerBindings in order to post/invoke V8 functions, and use
the new methods instead.

Notably, CreateClosureThatPostsV8Callback() was being used to wrap
everything into a base::Callback, but it introduced an extra PostTask()
in the process. We do not need that, so we use our new Callback-wrapper
methods directly instead of binding a PostTask in as well.

R=avi@chromium.org

Bug: 866140, 1069111
Change-Id: I09aaef11032c0a90162a3ad4b6800c958ce5f102
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199726
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768833}