weblayer: updates how BrowserControlsState is determined
This introduces BrowserControlsNavigationStateHandler, which is responsible
for determining the state of the controls as related to navigation state. It
has the logic of Clank and chromeos. The interesting changes are:
. the control state is specific to a navigation, and needs to be
  sent in DidFinishNavigation(). This is why prior to this patch the
  controls snapped back on every navigation (I'm not positive if the
  state is specific to a renderer process, or per-navigation,
  clank code says process, but I'm not sure that is right).
. The controls are forced visible until 3 seconds after loading.
. Lots of other logic around forcing shown around various error states.

I've put the logic in the C++ side for two reasons:
. We are striving to have C++ the source of truth.
. Tests are easier.

I would like to make this compile on all platforms, but for now it's
just android.

There are also some TODOs for parts we don't have wired up yet, but they
shouldn't stop this from landing.

I plan to write tests, but am trying to get this in before the branch.

I'm also disabling the bottom-tests, they are just too flaky. I know
the source of that flake and will have a patch to straighten them out
soon (it's not pretty).

BUG=1081231
TEST=forthcoming

Change-Id: Id1d7c36ed473b3df665e615372d2548d8e2bbcb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200318
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768835}