Allow finch-configured heap profiler collection interval
Default interval on iOS was set to 30 minutes, but it's unclear if
the value is ideal for iOS. Too frequent sampling will put unnecessary
load on the server, while to rare sampling will not allow us to
collect enough data.

Default interval on Android was set to 24 hours, which seems to be
too rare, so we don't collect enough data.

Configurable sampling interval will allow us to find the best sampling
rate for iOS and Android.

Bug: 1075702
TBR: rockot@chromium.org
Change-Id: I6e820486240e2b3552146d0b066d411b79a12a6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194593
Commit-Queue: Eugene But <eugenebut@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768838}