Add DCHECK that WebSockets have non-empty IsolationInfos.
Eventually we'll add similar DCHECKs in net, at the URLRequest layer,
but it's better to DCHECK sooner, with more relevant data on the stack,
and this also protects against regressions until we can DCHECK that all
URLRequests have an IsolationInfo.

Also fix a bug in DedicatedWorkerHost that the DCHECK caught.
There was an assumption that StartScriptLoad was called before other
methods that need an IsolationInfo, but this is not true.

Bug: 1082280
Change-Id: I14878ffa0e8e1fe4e7d2a6d4dba537acebd8c71c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196721
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768841}