Add a wrapper for the set of Stack animations
This wrapper wraps a AnimationSet and keeps track of the cancelable
animators, and it provides the ability to cancel those animation with
one function call. So we don't need to keep a list of FloatProperty and
iterate through the list to locate the cancelable animators.

Change-Id: I43956e790636365f791c8aea305678dfc8216e1a
Bug: 1070284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183483
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Mei Liang <meiliang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768852}