Remove set_owned_by_client() uses in tests
In these unit tests, the reason to not letting the views hierarchy to
own the view was due to the controller and delegate's shorter lifetime.
We changed to let the test class own the controller and delegate so that
their lifetime are guaranteed. Thus we can remove the use of
set_owned_by_client().

BUG=1044687

Change-Id: Id95274422abcbae29107ac2cfdc69adec4ba2d67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196929
Commit-Queue: Wei Li <weili@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768862}