Reland "blinkpy: Create WPTExpectationsUpdater subclass that handles Android"
import_notifier.py requires WPTExpectationsUpdater.write_test_expectations
to return a dictionary mapping test names to lists of test expectation
strings. This CL makes sure that it returns that dictionary.

The fix for WPT imports can be seen in the diff between patchset 1
and patchset 2.

Test: crrev.com/c/2200009 does a full run of wpt_import.py. Expectations
are created for the imported tests and are written to the file.

Bug: 1050760, 1050754
Change-Id: I138ab91408e52b450e47bf0b38b10158b5677c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199998
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: Luke Z <lpz@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Cr-Commit-Position: refs/heads/master@{#768882}