Update the V8 version for Chromium branch 84.
This is a roll-back to 8.4.371. This is the most stable candidate out of the latest Canaries.

If this commit results in failures in Blink please contact the Blink sheriff. They may need to make changes to Blink because of the roll-back.

When in doubt please try to contact the committer and reviewers of this CL before reverting it.

R=machenbach@chromium.org,vahl@chromium.org

Change-Id: Ieeda8c6c93c5746014cf213f2ca7ab4028f8783a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2201376
Reviewed-by: Michael Hablich <hablich@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768883}