[WebLayer] Remove urlbar textview's layout weight
A layout weight of 1 is a good way to make the TextView take up the rest
of the remaining width of the parent view, but it is causing
issues with centering of the view. This CL removes that.

It also centers the contents of the UrlBarView to match the current use
case.

Here's a video of what the UI looks like after this change:
https://drive.google.com/file/d/1fRzackboMhZW5mMAkiLyKs9zKbll35Ei/view?usp=sharing

Bug: 1052375, b/156453601
Change-Id: Iecf4b03aacc2ebb60253c6e437031de1320f644a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199277
Reviewed-by: Bo <boliu@chromium.org>
Auto-Submit: Mugdha Lakhani <nator@chromium.org>
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768892}