[Autofill] Added LINE1 / ZIP_OR_STATE to address requirements.
Relaxed requirements for BR to ZIP and for TR to ZIP_OR_STATE.

With this change, logging the import success is not conditioned
on having a LINE1 entry anymore. Since not needing a LINE1 entry
is a very rare case, it is not part of combined requirements
histogram, yet.

Change-Id: I2103a5b4f7d54216e5ea210faee03f91460d60fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199105
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768894}