Metal: Run test shaders on a ThreadPriority::DISPLAY thread
The Metal test shaders are compiled in a function that comes from
a PostTask.  There is currently no path through which a PostTask
will end up running on a ThreadPriority::DISPLAY thread.

Change the Metal test code to create a ThreadPriority::DISPLAY
thread and do its work there. A separate change is making that
correspond to a pthread priority of 50.

Of note is that Metal's callback to Chrome comes from a Metal-created
thread which has priority 31.

Also change the default test shader to be a Compile instead of a
Link (the Link tests have not been completed).

Bug: 974219
Change-Id: Ieb5ab0748b3aa255ee7dd3348870ddf812caa682
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200357
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768962}