Trust Tokens: Use a browser-provided feature policy on subresource loads
Trust Tokens (https://github.com/wicg/trust-token-api) is a new feature
allowing callers to specify cryptographic operations to execute
alongside certain requests. It currently provides an iframe interface
executing these operations alongside subframe navigations, and fetch
and XHR interfaces allowing executing the operations alongside
subresource requests.

Two of the three Trust Tokens operations ("redemption" and "signing,"
in contrast to "issuance") require that the trust-token-redemption
Feature Policy feature be present in the executing context.

This is the second of a series of two CLs enforcing this Feature Policy
requirement in the browser. It binds a frame's trust-token-redemption
feature policy value to its subresource loaders, so that the network
service can verify that subresource requests bearing redemption or
signing operations originated from frames with the necessary policy.

to the network service; adds NetworkContext unit tests confirming that
the checks work as intended given a value of the policy.

Test: Adds content browser tests confirming that the feature policy gets passed
Bug: 1071272
Change-Id: I2bfeff20ec44459f408b67cd48965757a6b987c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165645
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768971}