[AW][Dev-UI] add histogram to track crash state
This logs a histogram to check crash collection state.

This does a small refactor to move the work of creating the error
message into the callback. This is a small optimization (this work is
only needed if crash collection is disabled), but also makes it easier
to only log these cases when they're user facing.

Fixed: 1082506
Test: Verify the error UI looks correct in each state
Change-Id: Ib78bc5e788e7ec53a4ce591a3121ecdc8c17ec36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2200277
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#768983}