exo: Use panel rotation for drawing/reporting
panel_rotation() is the difference between the buffer and hardware's
scanlines as opposed to rotation() which is the "logical" uprightness
of the screen. This matters on devices where the logical orientation
does not match the hardware orientation.

There are two cases where exo uses the wrong rotation which means:
 - cursor images are drawn in the wrong orientation
 - the screen orientation is reported incorrectly to clients. This is
   ~fine~ until you start using applications which refuse to handle
   events which are "offscreen".

We model this CL after changes in crrev.com/c/1911344.

Bug: 1081085
Change-Id: Ia3a01b8dd5d745dbdb6a5b2220cae44061088a77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2195341
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#769038}