[Extensions UI] Wire up a feature switch for settings overridden dialogs
Introduce a feature switch, kExtensionSettingsOverriddenDialogs, to
control using a dialog-based UI instead of a bubble-based UI for the
post-install confirmations related to settings overridden by extensions.

Wire this up for use with the NTP overridden flow, set to display at the
same time the bubble would normally show. The two are mutually exclusive
and perform the same actions: the user can acknowledge the extension,disable the extension to go back to the old settings, or dismiss thedialog without making a choice (which will cause the dialog to re-show
in a future session).

Update the browser UI test for the NTP overridden dialog to exercise
this flow (rather than force-displaying the dialog).

Bug: 1079364
Change-Id: Ic222f0db666856f0d5baa7156bf7e49e7c4de482
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2196645
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769086}