[CrOS Settings] Remove nonexistent kDevice Subpage enum value
This was added by mistake since Device is a section, not a subpage. Note
that, despite the warning in the file, this feature has not yet launched
and it is still safe for me to change the numerical values for these
subpages.

Bug: 1080892
Change-Id: Idf2dc51710cbe65bcf84b42bae6b5f45aa78af8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202790
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769090}