Cros: Fix ripple ring cropping on hotseat edges
When long pressing icons, the ripple ring should crop to the hotseat
edges. Sometimes, the user will trigger the ink drop animation to show
while another ripple animation is finishing to hide. This will led into
the ring not being cropped properly on the edges.
This changes keeps a count of the buttons that need to crop the edges of
the hotseat so that one button doesn't deactivate rounded corners while
another active button needs them.
Added a test to catch regressions.

Bug: 1079330
Change-Id: I2c87e452a4c9ec633b4322312d8e596bf251a684
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2202964
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769091}