desktop-pwas: Add app icon shortcuts menu tests for extensions based PWA
This is a follow up to https://crrev.com/c/2146212 which added support
for App Icon Shortcuts Menu on Windows.

This CL improves test coverage for Extensions based PWAs with
Shortcuts Menu. Specifically, it exercises code that adds shortcut icon
data to the extensions manifest, writes icons to disk and installs the
PWA.

To do this, it adds manifest handlers for the
web_app_linked_shortcut_icons and web_app_shortcut_icons keys that were
added in the initial change. In the future, these handlers will be used
as part of BMO Migration and App Sync.

This CL also includes tests for these new manifest handlers.

Bug: 1076042
Change-Id: Ic570e3201018c32fa01a0b31d92dc0c13110c701
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2183713
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Chris Mumford <cmumford@google.com>
Cr-Commit-Position: refs/heads/master@{#769092}