Refactor password protection code to decouple reason and core logic.
This CL is purely for refactoring. It doesn't contain any functional
change.

Motivation:
1. Currently, reasons are set in many Is* and Can* functions. They
are delivered from here and there, which is hard to set them correctly.
(We made this error twice: https://crbug.com/1077420 and
https://crbug.com/1016540)
2. These bool functions don't follow single responsibility principle
and generate side effects(i.e. change the value of reason)

Benefits:
1. Decouple setting reason logic from the bool functions and create
a dedicated function named GetPingNotSentReason to set it in a single
place.
2. Simplified logic in the bool functions to make them more readable.
3. Reduce the number of parameters in several functions.
4. Avoid core logic depending on the value of reason. Use the actual
state of the pref/policy instead.

Bug: 1077420
Change-Id: I5acdee9421a3fdbafbe5d894a9dd6dc8efba2863
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2191454
Reviewed-by: Bettina Dea <bdea@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769118}