Update Ambient Authentication to be allowed for non-primary OTR profiles
Non-primary OTR profiles are not used to create browser windows and are
only technical means for a task that does not need to leave state after
it's completed.
Therefore they are removed from the control of incognito and guest mode
policies and ambient authentication is made allowed for them similar to
regular profiles.

Bug: 1074201
Change-Id: I397657e40b95d932305e3246fae6dca63dd8ff8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2199286
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769151}