Metal: Run test shaders on a ThreadPriority::DISPLAY thread
The Metal test shaders are compiled in a function that comes from
a PostTask.  There is currently no path through which a PostTask
will end up running on a ThreadPriority::DISPLAY thread.

Change the Metal test code to create a ThreadPriority::DISPLAY
thread and do its work there. A separate change is making that
correspond to a pthread priority of 50.

Of note is that Metal's callback to Chrome comes from a Metal-created
thread which has priority 31.

Also change the default test shader to be a Compile instead of a
Link (the Link tests have not been completed).

This is the second submission of this patch -- the first one had a
typo that caused it to always report the shader as not compiling.

TBR=avi

Bug: 974219
Change-Id: If17ef311e6e5d67304e5e56b04b5ea2fb803dcf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2203232
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769157}