Fix text run calculation for PDFs starting with a Unicode space
For presenting the text model of a PDF for accessibility we break the
PDF content into text runs. While breaking the text into text runs we
always start with the first non-space (Unicode) character. This
assumes that PDF content will not have a preceding space. However, it
is not always the case. We can see that this particular file
http://africau.edu/images/default/sample.pdf starts with a space.

The logic to calculate text run further goes on to create a text run
with a bound which has not taken care of the preceding space character.
When this information is passed on to PdfAccessibilityTree we report
that the text run starts from the space character but we report text run
bounds which is short by a space character. When we start arranging
characters in PdfAccessibilityTree we run short by the width of space
character. As a result we see an offset from the actual position of
characters.

The fix is to add the bounding box of the preceding space character in
the bounds of text run.

Bug: 1076752

Change-Id: Ic9b06307daf5dc3681a42ae8176288f5b3012886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167749
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ankit Kumar 🌪️ <ankk@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Virender Singh <virens@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#769180}